DTO refactoring
