+++ Darts_Score_Management/Controllers/GameRulesController.cs
@@ -1,6 +1,6 @@
 ﻿using Darts_Score_Management.CustomExceptions;
 using Darts_Score_Management.Data.Models;
-using Darts_Score_Management.DTOs.Game;
+using Darts_Score_Management.DTOs.Game.State;
 using Darts_Score_Management.DTOs.Throw;
 using Darts_Score_Management.Interfaces.ServiceInterfaces;
 using Microsoft.AspNetCore.Http;
