+++ Darts_Score_Management/DTOs/Game/Core/BaseGameDTO.cs
@@ -1,8 +1,8 @@
 ﻿using Darts_Score_Management.Enums;
 
-namespace Darts_Score_Management.DTOs.Game
+namespace Darts_Score_Management.DTOs.Game.Core
 {
-    public class GameSummaryDTO
+    public class BaseGameDTO
     {
         public int Id { get; set; }
         public GameType Type { get; set; }
@@ -10,8 +10,5 @@ namespace Darts_Score_Management.DTOs.Game
         public DateTime StartedAt { get; set; }
         public DateTime? EndedAt { get; set; }
         public bool IsComplete { get; set; }
-        public int PlayerCount { get; set; }
-        public int SetsCount { get; set; }
-        public int? WinnerId { get; set; }
     }
 }
