+++ Darts_Score_Management/DTOs/Game/Core/GameSettingsDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace Darts_Score_Management.DTOs.Game
+﻿namespace Darts_Score_Management.DTOs.Game.Core
 {
     public class GameSettingsDTO
     {
