+++ Darts_Score_Management/DTOs/Game/Response/GameResponseDTO.cs
@@ -1,6 +1,6 @@
-﻿namespace Darts_Score_Management.DTOs.Game
+﻿namespace Darts_Score_Management.DTOs.Game.Response
 {
-    public class GameListResponseDTO
+    public class GameResponseDTO
     {
         public int Id { get; set; }
         public string Type { get; set; }
@@ -9,7 +9,6 @@
         public DateTime StartedAt { get; set; }
         public DateTime? EndedAt { get; set; }
         public bool IsComplete { get; set; }
-        public List<GamePlayerInfoDTO> Players { get; set; } = new List<GamePlayerInfoDTO>();
         public int? WinnerId { get; set; }
     }
 }
