+++ Darts_Score_Management/DTOs/Game/State/GameStateDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace Darts_Score_Management.DTOs.Game
+﻿namespace Darts_Score_Management.DTOs.Game.State
 {
     public class GameStateDTO
     {
