+++ Darts_Score_Management/Interfaces/RepositoryInterfaces/IGameRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Darts_Score_Management.Data.Models;
-using Darts_Score_Management.DTOs.Game;
+using Darts_Score_Management.DTOs.Game.Response;
+using Darts_Score_Management.DTOs.Game.Statistics;
 
 namespace Darts_Score_Management.Interfaces.RepositoryInterfaces
 {
