+++ Darts_Score_Management/Interfaces/ServiceInterfaces/IGameService.cs
@@ -1,4 +1,6 @@
-﻿using Darts_Score_Management.DTOs.Game;
+﻿using Darts_Score_Management.DTOs.Game.Core;
+using Darts_Score_Management.DTOs.Game.Response;
+using Darts_Score_Management.DTOs.Game.Statistics;
 
 namespace Darts_Score_Management.Interfaces.ServiceInterfaces
 {
