+++ Darts_Score_Management/Repositories/GameRepository.cs
@@ -1,6 +1,8 @@
 ﻿using Darts_Score_Management.Data;
 using Darts_Score_Management.Data.Models;
-using Darts_Score_Management.DTOs.Game;
+using Darts_Score_Management.DTOs.Game.Players;
+using Darts_Score_Management.DTOs.Game.Response;
+using Darts_Score_Management.DTOs.Game.Statistics;
 using Darts_Score_Management.DTOs.Leg;
 using Darts_Score_Management.DTOs.Set;
 using Darts_Score_Management.DTOs.Throw;
